requirements.txt
opencv-python~=4.10.0.82
numpy~=1.26.4
onnxruntime~=1.18.0
pillow~=10.3.0
jinja2~=3.1.4
playwright~=1.44.0